Change service-type for nvidia.service
When changed to oneshot, the subsequent services wiill actually wait for the nvidia.service to finish, i.e. wait for drivers to be installed.
The subsequent services can be configured to wait for nvidia.service.
This needs to be coupled with RemainAfterExit=yes to ensure it doesn't get kicked-off automatically again.

https://github.com/flatcar/Flatcar/issues/1136